# openapi-generator/aspnetcore/common/src/Agl.Cdr.Common/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "contact@consumerdatastandards.gov.au",
      "name" : "Consumer Data Standards",
      "url" : "https://consumerdatastandards.gov.au"
    },
    "description" : "Consumer Data Standards APIs created by the Data Standards Body (DSB), with the Data Standards Chair as the decision maker to meet the needs of the Consumer Data Right",
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "CDR Common API",
    "version" : "1.15.0"
  },
  "servers" : [ {
    "url" : "https://data.holder.com.au/cds-au/v1"
  } ],
  "paths" : {
...
    "/discovery/status" : {
      "get" : {
        "description" : "Obtain a health check status for the implementation",
        "operationId" : "getStatus",
        "parameters" : [ {
          "description" : "Version of the API end point requested by the client. Must be set to a positive integer. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If the value of [x-min-v](#request-headers) is equal to or higher than the value of [x-v](#request-headers) then the [x-min-v](#request-headers) header should be treated as absent. If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable. See [HTTP Headers](#request-headers)",
          "in" : "header",
          "name" : "x-v",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Minimum version of the API end point requested by the client. Must be set to a positive integer if provided. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable.",
          "in" : "header",
          "name" : "x-min-v",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseCommonDiscoveryStatus"
                }
              }
            },
            "description" : "Success",
            "headers" : {
              "x-v" : {
                "description" : "The [version](#response-headers) of the API end point that the data holder has responded with.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseErrorListV2"
                }
              }
            },
            "description" : "The following error codes MUST be supported:<br/><ul class=\"error-code-list\"><li>[400 - Invalid Field](#error-400-field-invalid)</li><li>[400 - Invalid Version](#error-400-header-invalid-version)</li><li>[406 - Unsupported Version](#error-406-header-unsupported-version)</li></ul>"
          }
        },
        "summary" : "Get Status",
        "tags" : [ "Common", "Discovery" ],
        "x-version" : "1"
      }
    },
    "/discovery/outages" : {
      "get" : {
        "description" : "Obtain a list of scheduled outages for the implementation",
        "operationId" : "getOutages",
        "parameters" : [ {
          "description" : "Version of the API end point requested by the client. Must be set to a positive integer. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If the value of [x-min-v](#request-headers) is equal to or higher than the value of [x-v](#request-headers) then the [x-min-v](#request-headers) header should be treated as absent. If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable. See [HTTP Headers](#request-headers)",
          "in" : "header",
          "name" : "x-v",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Minimum version of the API end point requested by the client. Must be set to a positive integer if provided. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable.",
          "in" : "header",
          "name" : "x-min-v",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseDiscoveryOutagesList"
                }
              }
            },
            "description" : "Success",
            "headers" : {
              "x-v" : {
                "description" : "The [version](#response-headers) of the API end point that the data holder has responded with.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseErrorListV2"
                }
              }
            },
            "description" : "The following error codes MUST be supported:<br/><ul class=\"error-code-list\"><li>[400 - Invalid Field](#error-400-field-invalid)</li><li>[400 - Invalid Version](#error-400-header-invalid-version)</li><li>[406 - Unsupported Version](#error-406-header-unsupported-version)</li></ul>"
          }
        },
        "summary" : "Get Outages",
        "tags" : [ "Common", "Discovery" ],
        "x-version" : "1"
      }
    }
  },
  "components" : {
...
    "schemas" : {
      "ResponseCommonDiscoveryStatus" : {
        "example" : {
          "data" : {
            "updateTime" : "updateTime",
            "explanation" : "explanation",
            "expectedResolutionTime" : "expectedResolutionTime",
            "detectionTime" : "detectionTime",
            "status" : "OK"
          },
          "meta" : "{}",
          "links" : {
            "self" : "self"
          }
        },
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ResponseCommonDiscoveryStatus_data"
          },
          "links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "meta" : {
            "type" : "object"
          }
        },
        "required" : [ "data", "links" ],
        "type" : "object"
      },
      "ResponseDiscoveryOutagesList" : {
        "example" : {
          "data" : {
            "outages" : [ {
              "duration" : "duration",
              "outageTime" : "outageTime",
              "isPartial" : true,
              "explanation" : "explanation"
            }, {
              "duration" : "duration",
              "outageTime" : "outageTime",
              "isPartial" : true,
              "explanation" : "explanation"
            } ]
          },
          "meta" : "{}",
          "links" : {
            "self" : "self"
          }
        },
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ResponseDiscoveryOutagesList_data"
          },
          "links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "meta" : {
            "type" : "object"
          }
        },
        "required" : [ "data", "links" ],
        "type" : "object"
      },
      "DiscoveryOutage" : {
        "example" : {
          "duration" : "duration",
          "outageTime" : "outageTime",
          "isPartial" : true,
          "explanation" : "explanation"
        },
        "properties" : {
          "outageTime" : {
            "description" : "Date and time that the outage is scheduled to begin",
            "type" : "string",
            "x-cds-type" : "DateTimeString"
          },
          "duration" : {
            "description" : "Planned duration of the outage. Formatted according to [ISO 8601 Durations](https://en.wikipedia.org/wiki/ISO_8601#Durations) (excludes recurrence syntax)",
            "type" : "string",
            "x-cds-type" : "ExternalRef"
          },
          "isPartial" : {
            "description" : "Flag that indicates, if present and set to true, that the outage is only partial meaning that only a subset of normally available end points will be affected by the outage",
            "type" : "boolean",
            "x-cds-type" : "Boolean"
          },
          "explanation" : {
            "description" : "Provides an explanation of the current outage that can be displayed to an end customer",
            "type" : "string"
          }
        },
        "required" : [ "duration", "explanation", "outageTime" ],
        "type" : "object"
      },
...
      "Links" : {
        "example" : {
          "self" : "self"
        },
        "properties" : {
          "self" : {
            "description" : "Fully qualified link that generated the current response document",
            "type" : "string",
            "x-cds-type" : "URIString"
          }
        },
        "required" : [ "self" ],
        "type" : "object"
      },
...
      "MetaError" : {
        "description" : "Additional data for customised error codes",
        "properties" : {
          "urn" : {
            "description" : "The CDR error code URN which the application-specific error code extends. Mandatory if the error `code` is an application-specific error rather than a standardised error code.",
            "type" : "string"
          }
        },
        "type" : "object",
        "x-conditional" : [ "urn" ]
      },
      "ResponseErrorListV2" : {
        "properties" : {
          "errors" : {
            "description" : "",
            "items" : {
              "$ref" : "#/components/schemas/ResponseErrorListV2_errors"
            },
            "type" : "array"
          }
        },
        "required" : [ "errors" ],
        "type" : "object",
        "x-conditional" : [ "meta" ]
      },
      "ResponseCommonDiscoveryStatus_data" : {
        "example" : {
          "updateTime" : "updateTime",
          "explanation" : "explanation",
          "expectedResolutionTime" : "expectedResolutionTime",
          "detectionTime" : "detectionTime",
          "status" : "OK"
        },
        "properties" : {
          "status" : {
            "description" : "Enumeration with values. OK (implementation is fully functional). PARTIAL_FAILURE (one or more end points are unexpectedly unavailable). UNAVAILABLE (the full implementation is unexpectedly unavailable). SCHEDULED_OUTAGE (an advertised outage is in effect)",
            "enum" : [ "OK", "PARTIAL_FAILURE", "SCHEDULED_OUTAGE", "UNAVAILABLE" ],
            "type" : "string"
          },
          "explanation" : {
            "description" : "Provides an explanation of the current outage that can be displayed to an end customer. Mandatory if the status property is any value other than OK",
            "type" : "string"
          },
          "detectionTime" : {
            "description" : "The date and time that the current outage was detected. Should only be present if the status property is PARTIAL_FAILURE or UNAVAILABLE",
            "type" : "string",
            "x-cds-type" : "DateTimeString"
          },
          "expectedResolutionTime" : {
            "description" : "The date and time that full service is expected to resume (if known). Should not be present if the status property has a value of OK.",
            "type" : "string",
            "x-cds-type" : "DateTimeString"
          },
          "updateTime" : {
            "description" : "The date and time that this status was last updated by the Data Holder.",
            "type" : "string",
            "x-cds-type" : "DateTimeString"
          }
        },
        "required" : [ "status", "updateTime" ],
        "type" : "object",
        "x-conditional" : [ "explanation" ]
      },
      "ResponseDiscoveryOutagesList_data" : {
        "example" : {
          "outages" : [ {
            "duration" : "duration",
            "outageTime" : "outageTime",
            "isPartial" : true,
            "explanation" : "explanation"
          }, {
            "duration" : "duration",
            "outageTime" : "outageTime",
            "isPartial" : true,
            "explanation" : "explanation"
          } ]
        },
        "properties" : {
          "outages" : {
            "description" : "List of scheduled outages. Property is mandatory but may contain and empty list if no outages are scheduled",
            "items" : {
              "$ref" : "#/components/schemas/DiscoveryOutage"
            },
            "type" : "array"
          }
        },
        "required" : [ "outages" ],
        "type" : "object"
      },
...
      "ResponseErrorListV2_errors" : {
        "properties" : {
          "code" : {
            "description" : "The code of the error encountered. Where the error is specific to the respondent, an application-specific error code, expressed as a string value. If the error is application-specific, the URN code that the specific error extends must be provided in the meta object. Otherwise, the value is the error code URN.",
            "type" : "string"
          },
          "title" : {
            "description" : "A short, human-readable summary of the problem that MUST NOT change from occurrence to occurrence of the problem represented by the error code.",
            "type" : "string"
          },
          "detail" : {
            "description" : "A human-readable explanation specific to this occurrence of the problem.",
            "type" : "string"
          },
          "meta" : {
            "$ref" : "#/components/schemas/MetaError"
          }
        },
        "required" : [ "code", "detail", "title" ],
        "type" : "object"
      },
...
    }
  },
...
}